backend/pyproject.toml
[tool.poetry]
name = "istudiox-backend"
version = "0.1.0"
description = "AI app builder backend"
authors = ["Avadhut <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.30.0"
sqlmodel = "^0.0.21"
bcrypt = "^4.2.0"
python-jose = "^3.3.0"
requests = "^2.31.0"
httpx = "^0.27.0"
aiofiles = "^24.1.0"
python-multipart = "^0.0.9"
redis = "^5.0.1"
rq = "^1.16"
pydantic = "^2.7.0"
python-dotenv = "^1.0.1"
email-validator = "^2.1.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4"
isort = "^5.13"
mypy = "^1.10"
ruff = "^0.5.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
